+++ packages/types/lib/http.ts
@@ -1,0 +1,31 @@
+/**
+ * An object of HTTP headers.
+ */
+export type HTTPHeaders = Record<string, string | string[] | number | boolean | null>;
+
+/**
+ * Possible HTTP methods, as stolen from `axios`.
+ *
+ * @see https://npm.im/axios
+ */
+export type HTTPMethod =
+  | 'get'
+  | 'GET'
+  | 'delete'
+  | 'DELETE'
+  | 'head'
+  | 'HEAD'
+  | 'options'
+  | 'OPTIONS'
+  | 'post'
+  | 'POST'
+  | 'put'
+  | 'PUT'
+  | 'patch'
+  | 'PATCH'
+  | 'purge'
+  | 'PURGE'
+  | 'link'
+  | 'LINK'
+  | 'unlink'
+  | 'UNLINK';
